Add usage of serial port connection policy. Now ports will be filtered by whether they actually belong to LTO Flash! hardware. This can be disabled in the preferences via the 'Only connect to devices that can be identified as LTO Flash!' on the LTO Flash! tab.
**Mac**
This works on Mac by using IOKit to enumerate USB devices in the system that have the product and vendor IDs in LTO Flash! hardware's USB to serial chipset. Matching devices are then subjected to a second check. When the USB chip is programmed, the USB vendor and USB product names are set (Left Turn Only and LTO Flash! respectively). Only USB devices in which all four of these criteria match are considered valid LTO Flash! devices.

Finally, the way the serial ports in the system are named, the USB serial number in the hardware is used in the port name. So, USB ports whose serial numbers match with serial ports ultimately are the only ones that "match", and only those will be probed as LTO Flash! devices.

**Windows**
Windows xp vs. Vista and later use slightly different approaches based on the SetupDi native APIs.

The advantage of this approach is that, in a "device search" only ports that match these will be probed as actual LTO Flash! devices. Other serial ports will no longer need to be opened, listened for the LOCUTUS beacon, and then closed.