File Access has gone wild, and it's time to reign it in.
In this part of the performance tune-up, introduce the core FileMemo generic. Simple Memo pattern implementation that uses file existence, size, and last modification time to determine if data (memo) needs update.

To support this safely, bump the requirements up for the INTV.Core PCL. No more Windows Phone 8 support. :P It should work for 8.1, but didn't download the SDK, so it's just not going to work there. As if anyway. :P The bump in requirements is to use ConcurrentDictionary as the storage for the memos.

As a side effect of introducing FileMemo, do what should have been done long ago - make accessing the file system in INTV.Core via an interface rather than a manually provided list of individual delegates.

INTV.Shared implements the IStorageAccess interface and supplies it to INTV.Core.