CompareILfsFileInfo had some old code. Clean up comparison to use interfaces instead.
Exclude not only save fork, but manual fork for root file, which is exploited to store menu position data on the cart.

Instead of re-creating the IEnumerable of fork types to ignore during comparison, use static readonly arrays. You'd hope that the compiler would be smart enough to optimize that out -- it seems obvious -- but making it utterly clear can't hurt. In addition, cleaned up short name compares. This is so we can compare two FileSystems that are either both 'host' or 'device'. Previously, code assumed we'd only ever compare host vs. device.