Add 'SimpleCompare' which does a fast, less-accurate comparison of two FileSystem instances. It exits when the first difference is found.
There are two implementations, one of which is compiled out. The generic CompareTo uses'early exit' but the general infrastructure. Since it will accumluate a full 'difference' of some kind, it involves using the differences data structures, which includes adding an element to an enumerable data structure. Since checking to see if any of the difference accumulators have content is a bit more expensive, and the goal is speed, use an alternate implementation that never uses the difference accumulators.