NSApplication and NSApplicationDelegate were a bit messed up in MonoMac - as well as the usage and application of them in this code. Xamarin.Mac strongly enforces one-or-the-other. I.e. if you use event handlers -- which are a C# sneaky way to the delegate, then that's all you should use. If you use a delegate, then the delegate should be the only mechanism used.
The type for NSApplicationDelegate changed to strengthen this differentiation as well. Xamarin.Mac exposes this as an interface. If memory serves, using C# events spins up a pre-defined implementation of INSApplicationDelegate provided by Xamarin.Mac -- so rather than use both, switch to just using the delegate. This has the unfortunate result in putting some code into the app delegate -- which is not defined in this assembly.

This change attempts to keep things as-is for the MonoMac build.

And, of course, type fun. nint and friends.