Standard type tweaking stuff -- nint, nfloat, CGRect, CGSize, CGPoint.
Also, as in other code, the ability to edit columns in tables / trees is not well-exposed in the Xamarin APIs. At least for the controls in use here. The underlying Objective-C message function is also not in the standard ObjCRuntime in Xamarin.Mac, so p/Invoke it directly.

Happily, you can now correctly get the RepresentedObject from NSTreeNode in Xamarin.Mac, as well as assign the Content to NSTreeController. Actual Hack Reduction! When does that happen?!