Add helper method specifically for NSResponder-derived types to help ensure that event handlers run on the main thread. This will be sprinkled throughout the *Controller and a few *View classes to hopefully address some sporadic problems that can arise. Joe encountered a crash in power-cycling the console while LUI was attached and running that exposed one of these.
Only reason to hesitate somewhat on immediate widespread adoption is the possibility of this altering some behaviors. One in particular seems to be that the UI is not handling re-establishing communication with Locutus after launching a game and console long reset. The attempt to listen for beacon keeps timing out. But it recovers find after disconnect / reconnect within LUI.