Fix issue #20 - https://github.com/intvsteve/VINTage/issues/20
When adding a new ROM, we will combine features. If the new ROM is unknown, it would "combine" the 'Tolerates' setting (default for most) with 'Enhances' for video standard support - which means 'Unknown'. However, the result is that the ORing of those together produces a 'required' value for the video standard, which does not make sense.

To address existing behavior of saved ROM lists, coerce the read and and write operations on the Ntsc and Pal properties.

Also, fix how flags are combined to prevent this escalation from happening with other features. Unfortunately, if there are bugged ROM files out there for things like JLP and other features, they will remain bugged until the user manually fixes them, or deletes / re-adds the ROM after getting this fix.