You can now drop files into the ROM list!
Quick implementation of dropping files into the ROM list. This is a little bit icky, in that documentation / examples are ... sparse. Code does not do much in the way of validation w.r.t. validating the drag/drop data - it just jumps right into parsing it as an array of strings separated by newlines / carriage returns / or both.