Fix cosmetic issue in menu layout view
Got the stetic file in line with desire to have folder status expand and fill the area above the tree. That ensures that the buttons and combo box are pushed out to the right as intended.

A little bit of drag/drop code (event handler) has leaked in here too.