Fix add items to folder...
It appears that the 'put at front' code was always wrong - it was just never tested. The problem was that when the 'add items at index zero' code was adding the new item to the tree store, it used 'insert after' instead of 'insert before'. As a result, the tree store and the 'real' model (MenuLayout) became out of sync. This would happen whenever an 'insert before' operation would occur.