Fix a cute bug in which a missing file (manual, ROM, et. al.) can make it impossible to add new ROMs to the MenuLayout!
What was happening is that the behavior of the Fork(IRom) constructor was such that the Crc24 (UID) of the Fork as well as its Size (in bytes) would be uninitialized.

That worked in conjunction with how the GKT dedupes files. A missing fork's local file system file could cause that fork's Size and UID (Crc24) to be uninitialized.

The result was that an attempt to add "foo.rom" to the menu could "dedupe" to a fork for "baz.txt" - a missing manual file. The code would detect the missing fork file -- for the wrong fork -- and throw an exception, resulting in the ROM not being added!

The fix is subtle. The Fork.UpdateFilePath method will allow a full update in the case where the fork's ROM is not null. Chances for a regression here appear to be low.