Add stream InUse feature
IStreamConnections now have the ability to provide a scoped 'In Use' object. This simply marks the port as 'busy' in a generic way. The new InUse property can then be used by client code to take actions or prevent actions depending on that state.

In the wider sense, some icky code in the INTV.LtoFlash assembly could probably be substantially cleaner - mainly the interaction between device commands and the ping/garbage collect timer.

Also, this will provide a way to hook into power management to prevent system sleep during long-running commands, if such a feature is necessary.