Add ISystemPowerManagement
This interface will be used to provide a better way to hook into power events.

In reality, it's going to be used on Mac, where catastrophic crashes (kernel panics) have been known to happen if the system goes to sleep (low power mode) while the serial port is in use.

This will combine with the eariler 'InUse' feature to avoid those problems in a better way.