Implement ISystemPowerManagement
Use the new IOKit bindings to hook up to power management events from the OS. This means that the implementation of power management using NSWorkspace over in INTV.LtoFlash can be removed.

The NSWorkspace approach can still be used if we turn off the IOKit implementation for some reason (e.g. if it's buggy). Two years in, though, and it's not seeming tooooo buggy.