Add 'all' rule for Mac builds
The INTV.Version project on Mono is shared among three environments -- MonoDevelop (Linux/GTK), Xamarin Studio and Visual Studio for Mac. On the Mac platforms, it's necessary to also generate Info.plist. Hopefully this will do the trick. If not, it'll be time to check in Info.plist again, even though it *should* be generated.