Fix issue #203 and issue #204
Issue #203 is ECS.bin is not always passed to jzintv. LUI was only doing this when -s1 was needed based on ROM features and user settings. Change is to always send the -E /ecs/path/ECS.bin to jzintv since jzintv will only really care to load it if -s1 is also sent.

Custom command line setting in the case where this bug was found was including -s1 for all ROMs - which exposed the interaction between the two args.

For Issue #204, code had overlooked that settings data stores path strings as NSUrl data on Mac, so a 'resolve' is needed for the C# System.IO.Path validation code to work properly. Due to that oversight, hackfiles could *never* work on Mac because the validation code would never find files with paths that had URI syntax (file:///path/to/hackfile.txt).