Implement IProgramFIleMetadata for UnmergedProgramInformation
This is really just a stub implementation. This type is used to slip in extra ROM database entries that for one reason or another are not part of the INTV Funhouse ROM database.

A future project will create a true 'rominfo' database for INTV Funhouse, which should reduce (if not eliminate) the need for this.