Parse jlp_savegame data
This was being ignored. Initially it was introduced as a file name, but that was ill-conceived.  Database uses it to store number of flash sectors, though it's a string. Rather than alter the database schema, just parse the string as number of JLP flash data sectors.

Also tweaked usings a little bit.