Add helper for CRC matching of IRom using new ProgramIdentifier
When searching for a matching ROM via IProgramDescription's Rom field, we can do a more wide-ranging match and consider not only the overall file CRC, but also the originating CRC information embedded within LUIGI files. The new method also allows for optionally requiring CFG file CRC matches. If / when the day comes that CFG CRC is computed on a 'canonical' format of the CFG file, this could become quite powerful.