StreamUtilities overhaul
Changes brought about by writing unit tests. Unit test creation emphasized that the class was not testable, as when running xunit tests in parallel, the singleton nature of the registered storage access was a problem.

Practically speaking, this isn't much of an issue, as these are used as initialize once and move on.  It's all necessary for no good reason other than "hey, let's try using a Portable Class Library!" that needs to access a file system -- which PCLs don't offer direct access to. :P

Most important changes:
Don't expose as extension methods on string. Why did I do that anyway?
Renamed 'Size()' method to 'FileSize()' to better indicate what it does.
Support for providing an explicit IStorageAccess rather than the cached one. This is driven for testability purposes.

The fallout is lots of changes in other classes.

This commit will break a few things (various FileMemo implemetnations), which will be fixed subsequently.