Finish out WeakKeyDictionary implementation
It actually had an embarrassing number of incorrect and incomplete aspects to its implementation. Especially the enumerators. This came up when going for code coverage in tests... I.e. most of the exposed API surface is not used in the code -- at least in Windows.  Hopefully the fixed implementations don't totally hose the Mac build.