Buff code coverage for WeakKeyDictionary tests
It's as near 100% as it can be. There's a chance that some of these tests will be intermittent - particularly the ones that try to add values and then validate all of them.  Specifically, weak keyed dictionaries may have a tendency to ... "lose" values when the keys get GC'd. These tests currently run reliably on this dev's machine.