Tweaks to SPrintf in StringUtilities
Most of the changes are cosmetic, and help the Visual Studio 2012 code coverage tools more accurately report coverage. Really, it can't handle if (a && b) accurately? *sigh*

Some other changes are to disable or remove code that cannot be reached.

Fixed a couple of situations in which some combinations of format specifiers and data types would actually crash. These now are going to behave consistently, and not crash.