Added dummy ExcludeFromCodeCoverageAttribute
This is for Portable Class Library use only. Merely having an implementation of this attribute in the expected namespace suffices to disable code coverage. Conditional, private methods (e.g. various internal debugging facilities that are compiled out) shouldn't confuse the code coverage tools.

Sadly, anonymous methods (lambdas) that are in conditional methods that are to be ignored still show up as missing in code coverage. :P