Fix another intermittent WeakKeyDictionary test
Here we backing off validating result of CopyTo. The real trick here is that it's unpredictable exactly what will survive for the CopyTo operation! It's conceivable that *ALL* the entries in the weak key dictionary could be discarded between the population of the weak key dictionary and the execution of the CopyTo operation!