Remove an unused helper and make friendlier to code coverage analysis
Hopefully the removed overload of RaisePropertyChanged in the helpers isn't used on Mac.

Also: Rearranged some simple logic to better cope with code coverage tools. This is going to get really old, really fast ... I mean, incomplete coverage on:
if (a && b)
{
}
but, with the same test code, this gets a perfect score:
if (a)
{
  if (b)
  { ... }
}

Perhaps because if (a && b) doesn't actually test all four possible combinations?