Relocate and tighten up CFGVAR metadata parsing
The high-level loop that extracts all CFGVAR metadata "blocks" really should be on the CfgVarMetadataBlock class. Makes it more testable.

Also, ignore lines of text in a .CFG file until the parse hits a [vars] line - and resume ignoring lines if a different "section" is encountered.  The parser will honor multiple [vars] sections ... not sure if the parser for jzintv does the same.

Finally, added 'publishers' and 'versions' synonyms, as these have been in use by rom2bin, though that was in error. This will be rectified in an update to jzintv / SDK-1600, but since it's been in the wild, may as well cope with it.