Fixed a bug in metadata parser
It would return a null metadata block if it had a payload length of zero. However, an empty string metadata entry, e.g. a description, is perfectly valid, and has a zero payload length.

Also mostly implemented a TODO. If we hit a case where data CRC mismatches the computed CRC, tag the block as one to ignore.