Some tidying up of TestStorageAccess
Moved the fake FileSystem singleton into the storage access rather than leaving in the stream type.  Still a bit icky, as there are occasional race conditions when running tests that are related to the singleton nature of FileSystem.

Got rid of static methods, which were the source of some risk as well.  Tests should avoid modifying "files" in-place, but there are also likely problems with returning the same instance of a stream in which, when multiple tests run in parallel, muck with each other.  The stream is not synchronized, so ... ick.

Not addressing at this time though.