Add tests for LuigiFeatureFlagsHelpers and fix a bug!
While writing tests, found a subtle bug in the LuigiFeatureFlags enum... It sure appears in the debugger that 1<<32 wraps the bit back around, which I did not expect.  Did not determine if the problem was due to the shift being inside a ulong enum or what... but the test uncovered this latent bug, so yay!