Enable actual 'tertiary' support for getting IProgramInforamtion
This is done in the context of writing tests.  Perhaps we're going to be calling intvname too often now, which may become a problem, in which case the three-layered approach must be reexamined.  As it stands, tests will now be checking that GetProgramInformation uses all the levels as expected.

Also, an embarrassing bug in GetProgramMetadata, which was never going to return metadata for RomFormatRom files!