Add IRomHelpers.GetProgramInformation() tests
In the course of getting code coverage via these tests, fully implemented the three-tier system for merging program information (hopefully!) as well as adjusting how ProgramFeatures are combined.  ProgramFeatures.Combine is a bit more intelligent now, in that it will override default values with non-default values when combining features in most cases.