Fix documentation and prevent unlikely bug
Documentation for EnsureCfgFileProvided() was pretty wrong - specifically the return value documentation.

Also, avoid a potential (but zero percent chance of happening in actual product) bug in which systems that formatted integral numbers converted to stings to use grouping symbol (, or .) would produce bogus file name for stock cfg files.  Given there are only 10 (named 0-9) this could not happen in practice. But an ounce of prevention and all that.

Finally, removed unreachable code in EnsureCfgFileProvided.