Properly deal with bogus date from intvname
Moving to the new date/time builder and populating metadata, real-world testing (to this point only fakes for intvname are being used in tests) reveals a behavior of intvname that is unfortunate.  For scrambled LUIGI ROMs, it returns a year of "-1" as a string.  The program info implementation was not guarding against garbage from intvname, and did what it should - the ArgumentOutOfRangeException was thrown creating the release date metadata.  This has been fixed.

Also of note: Tests for UserProgramInformation have not yet been written. Hence, this wasn't found previously.