Add JlpFeaturesHelpers.MaxTheoreticalJlpFlashSectorUsage
This constant is just a better named version of the theoretical maximum number of JLP Flash save data sectors that can be used.  In practicality, the value is much lower on JLP hardware, though it could be much larger on LTO Flash! boards or possible future revisions of the JLP hardware.