Bug fixes and finishing out a few things in IntvFunhouseXmlProgramInformation
Bug fix:
The parsing for 'incompatibilities' was done wrong.
The bug has been masked by the fact that 'Incompatibilities' is an unused feature in the current INTV Funhouse database.  It was speculatively introduced and subsequent reworking of the features system in the various ROM metadata systems has likely rendered it obsolete.  But, until that's been verified, we should at least make the code behave correctly.

Finishing out:
The 'Additional Information' metadata has been augmented to take advantage the 'vendor' and 'original vendor' data in the database by adding the vendor website URLs to the additional information metadata.

Finally, some cosmetic rearrangements to reduce confusion in code coverage, and also to eliminate guarding against unreachable code branches.