Fix oddly partial ProgramInformationBuilder.WithYear implementation
It was only setting years at or after 1978.  This makes a certain kind of sense given "reality" but belongs in a validation type function instead.  This will likely be revisited when unit tests are written.

The implementation now enforces ranges that are supported by ROM metadata, which is still somewhat arbitrary.