Fix some tests that expect different behaviors in DEBUG vs RELEASE
Some code that throws exceptions in debug does not do so in release. These changes allow the tests to succeed in debug or release.

Also of note: Code coverage has some bugs in release builds, specifically around switch statements. Apparently the IL generated in release builds can result in unreachable code, depending on how the code is generated.  This is a fault of the tooling, not the tests.