Try again to fix an intermittent failing test
In RELEASE builds, this test failed with a relatively high degree of intermittency.  Did not rigorously investigate, but suspect it's due to immutable string type being GCd or something.  Created a class that boxes an int and had better success.  Odd that the intermittency was consistently for this test ... but re-running usually passed.