Clean up ProgramFileKindHelpers and add tests
The age and prototype nature of this from the pre-LTO Flash! days is clear in this code.  Fixed up some horrid stuff - public static fields that were modifiable dictionaries.  Eek! Exposed Lists, too.  These are left over from hack-n-slash refactor that pulled part of the helpers out into the INTV.Shared assembly.  Or was it that part of it was pulled out of INTV.Shared down into INTV.Core?)

In any case, this is tidier now, and has unit and code coverage tests.