Cleanup for testing ProgramSupportFiles
Remove exclusion from code coverage for private constructor.  Tests will use XML fragments to cover it.

Buff some documentation.

Attempted to remove direct exposure of List<> but it's necessary due to XML serialization.  Ideal solution would be to split out a separate interface and mark this class as 'restricted' -- public so XML serializer can create it, but update usages to use the interface instead.  Not doing that at this time.

Due to how the type is implemented, removed a null-check of AlternateRomImagePaths, as it's never going to be null.

Added a null-check of program description in the validation method.  Validation done with the ROM referred to in the instance is just as valid as using the ROM instance in the provided description.

Added a long TODO comment regarding validation of scrambled ROMs.  Perhaps there is nothing TODO, but at least now the thinking behind the behavior of validating scrambled ROMs when LTO Flash! is not connected is written down.

Added a TODO for validating file kinds other than ROM.  It may never happen.

Added comment and TODO about who the behavior of setting a support file value (a default one) to NULL just removes it from its corresponding 'alternates' which effectively promotes the first alternate to be the new default.  This is a bad API design.  Setting should "set" not have this subtle behavior.   This is another area that would benefit from splitting out the XML serialization type from the type exposed for use by the rest of the code.