Fix corruption found with scrambled LUIGI file
A user discovered that, due to weaknesses in how the UI was searching for program metadata in scrambled LUIGI ROMs, it was possible to corrupt the ROMs.xml as well as MenuLayout.xml files.

The changes leading up to this one have been focused on more strongly validating the contents of various LUIGI data blocks by strictly checking for corruption via CRC checks built into the file format.

This fix caps that process by ensuring that, once a scramble block is hit, that no further attempts are made to decode blocks in a LUIGI file.

The bug, in this case, was that by chance, scrambled data immediately following the scramble key block had a byte pattern similar to a LUIGI metadata block.  Basic CRC checks would have prevented that from being parsed - but instead they were skipped and parsing proceeded as if the data were valid.  This led to binary data being stored in the program name.  Subsequently, parsing failed.