Add helper to get first and last quotation mark characters
This method operates on the raw bytes in a buffer presumed to have its origins as a string.  As such, it does a simplistic search for the first and last quote characters by treating each byte individually.  *Hopefully* there are not any situations in which a raw byte that is part of a UTF-8 encoded UNICODE string is the first or last byte.  This shouldn't be a problem for the first quote, but is maybe, possibly, a thing for a character that is UTF-8 encoded in multiple bytes.