Add support for alternate MSYS path
The MSYS2 version of svn (1.11 as of this submission) hangs when running svn export of the repo.  In fact, this happens even with older versions of svn in MSYS2.  The hang appears to be in msys-crypto-1.1.

To avoid this, support an alternative location of svn via MSYS.

This is only honored and used bin the Installer.source batch file builder in Windows.  Older versions of svn appear to work just fine in Windows.   Mac build is unaffected, but at the same time, it, too, is likely running an older version of svn.

If it is the case that it's svn-version-specific, then this needs to be addressed in a different way eventually.