When building installer, report error accurately if files are dirty
The error messages was using the SVN_REPO value to report that there were dirty files, even when building using GitHub sources.

Now, if the SVN_REPO is empty, use the GIT_REPO name instead.  Since this house of cards only supports one repo at a time, this fix is good.  It now tells you the name of the repo that's got dirty files now instead of reporting a name of ''.