Remove unescape and bad character check from ROM credits metadata
The unescape code should not be present.

The 'bad character' check is to be addressed (eventually) in a different way.

As it stands, bogus UTF-8 data will be decoded as 'unknown / invalid' UNICODE characters.  "Invalid" characters such as control characters are not to cause rejection of the value.   Instead, if they are present, they will be retained.   They are valid UNICODE characters.  If they are to be stored in XML, they will be encoded adequately in a subsequent change.

Updated tests accordingly.