Fix extraction of embedded firmware file to disk
Found a bug in which the embedded firmware image would not be extracted to the file system if an older version had already been extracted.

The cause is that the resource was always being extracted to LTO_Flash_FW.upg.  If that file already existed, regardless of version, the resource would not be extracted.

Now, using a new feature in INTV.Shared, the firmware version will be appended to the file name.

The existing code already inspected existing files in the FirmwareUpdates directory to scrounge for updates.  Now it should put the new one there if it has one!