Add DeviceActivity enumeration
This is part of fixing issue #129. Specifically, the mechanism by which 'on connect' actions that are also controlled by a user preference are executed is pretty ad-hoc and buggy.

This enum simply formalizes the list of these actions that we currently care about so a system can be introduced to more methodically service them.