Add mechanism to register 'device activity' handlers
The Device class now support the ability to register a delegate to call when a specific DeviceActivity occurs.  This is meant to be used to address issue #129, and can be extended if other similar types of activities need to be added in the future.

During the timer callback that monitors Locutus device state, all registered activity handlers are call.  These handlers are responsible for determining whether they can run, and whether to continue running.

These delegates are all called from the main application thread, as to this point, they tend to desire to execute user-facing actions such as checking file system state or prompting the user to update device firmware.