Add EnableOnboardConfigMenu configuration bit
Not yet exposed in the UI. This was "leaked" in Firmware version 4416, and is part of a set of features for National Videogame Museum.

NOTE: This is LEAVING the reserved bit set, instead of changing reserved bits to 0xFF80000000000000, as the feature is not yet supported.

This *may* not be the desired behavior at this time.  Eventually, the intent is to stop mixing the reserved bits back in, so that if FW is ahead of the UI, the UI wont stomp the existing FW settings from the cart.