Fixed corner, corner case for ECS compatibility flags setting
Unit test coverage investigation discovered that the code was written such that it was not properly detecting 'all flags' from the device. This happened because it was stripping out reserved bits - making the 'all flags set' check impossible.  As a result, if 'all flags' were set for ECS flags, we defaulted to disabled instead of the intended default.

This is an extreme case since typically because these bits are typically never in this state - even from the factory.