Preserve reserved raw status bits from low 64
We explicitly break out specific groupings of status bits for clarity and convenience in the Device model. However, at this moment, we are blindly assuming absolute accuracy and knowledge about which bits are 'reserved' and which are not.

In fact, with out-of-band firmware updates, it is possible that the firmware uses configuration bits differently than what this code assumes.  Therefore, preserve bits received from hardware -- that the Model considers 'reserved' AS-IS --  without modification.

Subsequent changes will mix these bits AS-IS back into Set Configuration commands sent to the device in order to avoid inadvertently modifying the device configuration.