Add RandomizeJlpRam to ViewModel
This is the logical inverse of Device.ZeroJlpRam. The ViewModel takes care of the inversion.  Not sure if that's a best practice, but as we bridge between View and Model, rather than implementing the transform between the ViewModel and the View, per platform, just do the easiest thing and have the ViewModel handle te logic directly.