SetRandomizeLtoFlashRamCommand now firmware-aware
To make the 'Randomize LTO Flash RAM' command firmware-sensitive, add the configuration bits to the command data. Due to the more subtle nature of the enabled state of the command, use a more traditional command binding in the XAML snippet - and make the tool top description a one-way binding rather than one-time binding.