Fix launch crash in Xamarin.Mac build
When making RelayCommandBase derive from NSObject in order to support KVC binding, it turns out that *something* about MEF instantiation of a type happens "too early" - in Xamarin.Mac 5.8.0.0 / Mono 5.18.1.3. (This problem does *NOT* happen in the old MonoMac version (which is also 32-bit). The result is that the (default) NSObject constructor crashes when the static member of Intellicart's settingsPageViewModel is initialized.

This actually points to something ... icky ... about the data types LUI is importing very early in its process init - or a substantially different behavior for MEF in newer Mono / Xamarin.Mac release. I.e. why are the settings page view models being imported so early?

Anyhow, the crash avoidance is to lazily create the command rather than just creating it as a static field. This appears to work. An alternative may be to relocate the command to the command group and just have  the dialog reference the command that way. This was simple enough to implement.