Call command's OnCanExecute prior to considering app busy to block
Logically, this should not allow any command to execute when it should not. However, it does allow the CanExecute handler to run - even when the app is "busy". The logic behind this is that a 'CanExecute' may have side effects, such as changing tip strips / icons / etc. *in addition to* causing command visuals to be disabled.