[Mac] More cleanup in device info dialog
Remove need for backing fields for Boolean settings. Not sure why they
were there, but with new configurable features, perhaps the need is gone.
This simplifies the code.

Another cleanup that slipped in is how the 'block when app busy' is
maintained.  There was this odd 'only change the behavior when the
active page in the dialog is changed' code... Which of course would
run *every* time a page was changed, and change the setting for all
the affected commands anyway.  And the setting was never restored
when the dialog was closed.  Perhaps at this time the affected commands
are only displayed in the device info dialog, but if that changes, it would
be better for other visuals to do their own thing rather than have this
leak out to them, leaving behind a permanently altered behavior for
the command.