[All] Remove early raising of value change event
For configurable features, this is a subtle, but big change.
When setting the value on the model, do not immediately
update the value of the feature.  Instead, just execute the
command on the device and let the result come back via the
eventual status update reply.  The thinking behind this change
is that immediately raising value change even though the
hardware has not confirmed the change causes strange pulses
in the UI.  E.g. if there is an in-flight change, the UI updates
its displayed value immedately, but between the execution
of the command to change the setting (which happens off
the UI thread) and the actual change on hardware, it's possible
to get a result back from the background poll that has the
previous value.  The result is that the UI goes back to the "old"
value ... and then goes to the "new" value once the ping loop
closes.

Perhaps one *could* update the current value upon successful
command execution and more immediately raise a property
changed event in the model.  That could happen with a tighter
integration between execution of set configuration and the
property code itself, tightening the potential 'lag' between the
UI and actual hardware state down to a max of ~1.2 seconds.

Note that this change will also break some unit tests, which
will be updated in a subsequent commit.