[Mac] Device settings page shows device state...
when quickly changing multiple settings. That is, quickly changing settings
in succession can create the illusion that a setting has changed in the UI,
but subsequent updates will 'override' it back to its previous value.  The UI
was not displaying this behavior accurately because it was only raising change
events for the changed property from the perspective of hardware. This
arises as follows:
1. Property change A sends request to update flags
2. Property change B sends request to update flags

What occurs is that, in rapid succession, the flags from 'A' are sent, but before
receiving confimation from hardware, the flags for 'B' are also sent - without
the update flags from 'A' - thus stomping 'A'.

This is fallout from removing the state from the Device model. Addressing
the notion of 'pending' settings is something to address in subsequent
updates to how the configurable features implementation works.