Initial .NET 4.0 ZipArchive support
This is a pretty hacky implementation, as it uses Reflection to call the internal Microsoft ZipArchive API that's squirreled away in .NET 4.0.

Each API has been hand-tested with a small number of cases:
- Enumerate files in .zip
- Create a clone of a .zip
- Remove a file from a .zip

This was done by hand, and running in the debugger on a Windows 10 machine. Whether this works in a release, installed build in real Windows xp is TBD.